Исправлен баг, из-за которого перманентные варны снимались почти сразу

+++ src/main/java/workers/WarnWorker.java
@@ -12,7 +12,7 @@ public class WarnWorker implements Worker {
     public void execute() {
         Database db = new Database();
         for (Warn warn: db.getWarns()) {
-            if (warn.getEndTime() <= new Date().getTime()) {
+            if (warn.getEndTime() <= new Date().getTime() && warn.getEndTime() != 0) {
                 db.deleteWarn(warn.getGuildID(), warn.getWarnID());
             }
         }
